let's see how good ai is at refactoring
