+++ resources/weather/src/shared/biome.ts
@@ -42,6 +42,30 @@ export enum WeatherType {
   WHITEOUT = 'WHITEOUT',
 }
 
+export const WeatherHashes: Record<WeatherType, number> = {
+  [WeatherType.BLIZZARD]: GetHashKey('blizzard'),
+  [WeatherType.CLOUDS]: GetHashKey('clouds'),
+  [WeatherType.DRIZZLE]: GetHashKey('drizzle'),
+  [WeatherType.FOG]: GetHashKey('fog'),
+  [WeatherType.GROUNDBLIZZARD]: GetHashKey('groundblizzard'),
+  [WeatherType.HAIL]: GetHashKey('hail'),
+  [WeatherType.HIGHPRESSURE]: GetHashKey('highpressure'),
+  [WeatherType.HURRICANE]: GetHashKey('hurricane'),
+  [WeatherType.MISTY]: GetHashKey('misty'),
+  [WeatherType.OVERCAST]: GetHashKey('overcast'),
+  [WeatherType.OVERCASTDARK]: GetHashKey('overcastdark'),
+  [WeatherType.RAIN]: GetHashKey('rain'),
+  [WeatherType.SANDSTORM]: GetHashKey('sandstorm'),
+  [WeatherType.SHOWER]: GetHashKey('shower'),
+  [WeatherType.SLEET]: GetHashKey('sleet'),
+  [WeatherType.SNOW]: GetHashKey('snow'),
+  [WeatherType.SNOWLIGHT]: GetHashKey('snowlight'),
+  [WeatherType.SUNNY]: GetHashKey('sunny'),
+  [WeatherType.THUNDER]: GetHashKey('thunder'),
+  [WeatherType.THUNDERSTORM]: GetHashKey('thunderstorm'),
+  [WeatherType.WHITEOUT]: GetHashKey('whiteout'),
+};
+
 // All available weather variants (used as fallback)
 export const WeatherVariants: Record<WeatherType, string[]> = {
   [WeatherType.BLIZZARD]: ["BLIZZARD_winter2"],
@@ -182,6 +206,24 @@ export const BiomeWeatherVariants: Record<BiomeType, Partial<Record<WeatherType,
     [WeatherType.DRIZZLE]: ["DRIZZLE_finale1"],
     [WeatherType.SHOWER]: ["SHOWER_finale2"],
   },
+};
+
+/**
+ * Human-readable biome names for display
+ */
+export const BiomeNames: Record<BiomeType, string> = {
+  [BiomeType.GRIZZLIES]: 'Grizzly Mountains',
+  [BiomeType.TALL_TREES]: 'Tall Trees',
+  [BiomeType.BIG_VALLEY]: 'Big Valley',
+  [BiomeType.HEARTLANDS]: 'The Heartlands',
+  [BiomeType.GREAT_PLAINS]: 'Great Plains',
+  [BiomeType.BAYOU]: 'Bayou Nwa',
+  [BiomeType.LEMOYNE]: 'Lemoyne',
+  [BiomeType.NEW_AUSTIN]: 'New Austin',
+  [BiomeType.RIO_BRAVO]: 'Rio Bravo',
+  [BiomeType.ROANOKE]: 'Roanoke Ridge',
+  [BiomeType.CUMBERLAND]: 'Cumberland Forest',
+  [BiomeType.SCARLETT]: 'Scarlett Meadows',
 };
 
 // Define which weather types are allowed in each biome
